requirements.txt
tensorflow==2.19.0
gradio==6.0.0
pillow
numpy
